requirements.txt
torchvision

pyyaml

numpy

scikit-learn

matplotlib

pandas

scipy

h5py

